Yonas workspace

+++ src/main/java/CS544/Controller/ReactionController.java
@@ -1,14 +1,17 @@
 package CS544.Controller;
 
-import CS544.Dao.IReactionDao;
+import CS544.Model.Post;
 import CS544.Model.Reaction;
+import CS544.Model.User;
+import CS544.Service.PostService;
 import CS544.Service.ReactionService;
-import jakarta.annotation.Resource;
+import CS544.Service.UserService;
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.view.RedirectView;
 
 import java.util.List;
 
@@ -18,6 +21,12 @@ public class ReactionController {
 
     @Autowired  // @Resource
     private ReactionService reactionService;
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private PostService postService;
 
     @GetMapping("/")
     public List<Reaction> getAll(){
@@ -29,22 +38,51 @@ public class ReactionController {
         return reactionService.getReaction(id);
     }
 
-    @PostMapping(value = "/add")             // add reaction
-    public RedirectView addReaction(@Valid @RequestBody Reaction reaction){
-        reactionService.saveReaction(reaction);
-        return new RedirectView("/reaction/" + reaction.getId());
+    @GetMapping(value = "/getPost/{post_id}", produces = "application/json")
+    public List<Reaction> getPost(@PathVariable long post_id){
+            return reactionService.getByPostId(post_id);
     }
 
-    @PutMapping(value = "/add/{id}")
-    public void put(@PathVariable long id, @Valid @RequestBody Reaction reaction){
-        if (id != reaction.getId()) {
-            throw new IllegalArgumentException();
+    @PostMapping(value = "/add/post/{post_id}")
+    public ResponseEntity<Reaction> addReaction(@Valid @RequestBody Reaction reaction, HttpServletRequest request, @PathVariable Long post_id) {
+
+        Claims claims = (Claims) request.getAttribute("claims");
+        String username = claims.getSubject();
+        User user = userService.findByUserName(username);
+        Post post = postService.get(post_id);
+
+        if (user == null || post == null) {
+            throw new IllegalArgumentException("Invalid User ID or Post ID");
         }
-        reactionService.updateReaction(reaction);
+
+        else {
+            reaction.setPost(post);
+            reaction.setReactor(user);
+            reactionService.saveReaction(reaction);
+            return ResponseEntity.ok(reaction);
+        }
+    }
+
+    @PutMapping(value = "/update/{id}")
+    public ResponseEntity<Reaction> put(@PathVariable long id, @Valid @RequestBody Reaction reaction){
+        Reaction reaction1 = reactionService.getReaction(id);
+        if(reaction1==null){
+            return ResponseEntity.notFound().build();
+        }
+        reaction1.setReacted(reaction.getReacted());
+        reactionService.updateReaction(reaction1);
+        return ResponseEntity.ok(reaction1);
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public void delete(@PathVariable long id){
+
+    public ResponseEntity<String> delete(@PathVariable long id){
+        Reaction reaction1 = reactionService.getReaction(id);
+        if(reaction1==null){
+            return ResponseEntity.notFound().build();
+        }
         reactionService.deleteReaction(id);
+        return ResponseEntity.ok("Reaction updated successfully!!");
     }
+
 }
